+++ src/common/fees/makeFees.ts
@@ -1,5 +1,5 @@
 import * as bs from 'biggystring'
-import { asMaybe } from 'cleaners'
+import { asMaybe, Cleaner } from 'cleaners'
 import { Disklet } from 'disklet'
 import {
   EdgeIo,
@@ -11,13 +11,8 @@ import { makeMemlet, Memlet } from 'memlet'
 
 import { removeUndefined } from '../../util/filterUndefined'
 import { FEES_PATH, INFO_SERVER_URI } from '../constants'
-import {
-  asSimpleFeeSettings,
-  PluginInfo,
-  SimpleFeeSettings
-} from '../plugin/types'
+import { asFeeInfo, FeeInfo, PluginInfo } from '../plugin/types'
 import { calcMinerFeePerByte } from './calcMinerFeePerByte'
-import { processInfoServerFees } from './processInfoServerFees'
 import { processMempoolSpaceFees } from './processMempoolSpaceFees'
 
 interface MakeFeesConfig extends Common {
@@ -35,7 +30,7 @@ export interface Fees {
   stop: () => void
   clearCache: () => Promise<void>
   getRate: (edgeSpendInfo: EdgeSpendInfo) => Promise<string>
-  fees: SimpleFeeSettings
+  feeInfo: FeeInfo
 }
 
 export const makeFees = async (config: MakeFeesConfig): Promise<Fees> => {
@@ -43,9 +38,9 @@ export const makeFees = async (config: MakeFeesConfig): Promise<Fees> => {
   const { currencyInfo, engineInfo } = pluginInfo
 
   const memlet = makeMemlet(disklet)
-  const fees: SimpleFeeSettings = await fetchCachedFees(
+  const feeInfo: FeeInfo = await fetchCachedFees(
     memlet,
-    engineInfo.simpleFeeSettings
+    engineInfo.defaultFeeInfo
   )
   // The last time the fees were updated
   let timestamp = 0
@@ -59,10 +54,10 @@ export const makeFees = async (config: MakeFeesConfig): Promise<Fees> => {
       mempoolSpaceFeeInfoServer: engineInfo.mempoolSpaceFeeInfoServer
     })
     const cleanedVendorFees = removeUndefined(vendorFees ?? {})
-    Object.assign(fees, cleanedVendorFees)
+    Object.assign(feeInfo, cleanedVendorFees)
     timestamp = Date.now()
 
-    await cacheFees(memlet, fees)
+    await cacheFees(memlet, feeInfo)
   }
 
   return {
@@ -70,9 +65,9 @@ export const makeFees = async (config: MakeFeesConfig): Promise<Fees> => {
       const edgeFees = await fetchFees({
         ...common,
         uri: `${INFO_SERVER_URI}/v1/networkFees/${currencyInfo.pluginId}`,
-        processor: processInfoServerFees
+        cleaner: asMaybe(asFeeInfo, null)
       })
-      Object.assign(fees, edgeFees)
+      Object.assign(feeInfo, edgeFees)
       await updateVendorFees()
       vendorIntervalId = setInterval(
         // eslint-disable-next-line @typescript-eslint/no-misused-promises
@@ -109,32 +104,30 @@ export const makeFees = async (config: MakeFeesConfig): Promise<Fees> => {
 
       const rate = calcMinerFeePerByte(
         sumSpendTargets(spendTargets),
-        fees,
+        feeInfo,
         networkFeeOption,
         customNetworkFee[customFeeTemplate.key]
       )
       return rate
     },
 
-    get fees(): SimpleFeeSettings {
-      return fees
+    get feeInfo(): FeeInfo {
+      return feeInfo
     }
   }
 }
 
 const fetchCachedFees = async (
   memlet: Memlet,
-  fallback: SimpleFeeSettings
-): Promise<SimpleFeeSettings> => {
+  fallback: FeeInfo
+): Promise<FeeInfo> => {
   const data = await memlet.getJson(FEES_PATH).catch(() => undefined)
-  const feeSettings = asMaybe(asSimpleFeeSettings, fallback)(data)
+  const feeSettings = asMaybe(asFeeInfo, fallback)(data)
   return feeSettings
 }
 
-const cacheFees = async (
-  memlet: Memlet,
-  fees: SimpleFeeSettings
-): Promise<void> => await memlet.setJson(FEES_PATH, fees)
+const cacheFees = async (memlet: Memlet, feeInfo: FeeInfo): Promise<void> =>
+  await memlet.setJson(FEES_PATH, feeInfo)
 
 const sumSpendTargets = (spendTargets: EdgeSpendTarget[]): string =>
   spendTargets.reduce((amount, { nativeAmount }) => {
@@ -145,17 +138,17 @@ const sumSpendTargets = (spendTargets: EdgeSpendTarget[]): string =>
 
 interface FetchFeesArgs<T> extends Common {
   uri: string
-  processor: (fees: unknown) => T
+  cleaner: Cleaner<T>
 }
 const fetchFees = async <T>(args: FetchFeesArgs<T>): Promise<T | null> => {
-  const { uri, processor, io, log } = args
+  const { uri, cleaner, io, log } = args
 
   try {
     const response = await io.fetch(uri)
     if (!response.ok) throw new Error(`Error fetching fees from ${uri}`)
 
     const fees = await response.json()
-    return processor(fees)
+    return cleaner(fees)
   } catch (err) {
     log(err.message)
     return null
@@ -168,12 +161,12 @@ interface FetchFeesFromVendorArgs extends Common {
 
 const fetchFeesFromVendor = async (
   args: FetchFeesFromVendorArgs
-): Promise<Partial<SimpleFeeSettings>> => {
+): Promise<Partial<FeeInfo>> => {
   if (args.mempoolSpaceFeeInfoServer != null) {
     const mempoolFees = await fetchFees({
       ...args,
       uri: args.mempoolSpaceFeeInfoServer,
-      processor: processMempoolSpaceFees
+      cleaner: processMempoolSpaceFees
     })
     if (mempoolFees != null) {
       return mempoolFees