+++ src/common/utxobased/info/bitcointestnet.ts
@@ -53,7 +53,7 @@ export const engineInfo: EngineInfo = {
   defaultFee: 1000,
   feeUpdateInterval: 60000,
   mempoolSpaceFeeInfoServer: 'https://mempool.space/api/v1/fees/recommended',
-  simpleFeeSettings: {
+  defaultFeeInfo: {
     lowFeeFudgeFactor: undefined,
     standardFeeLowFudgeFactor: undefined,
     standardFeeHighFudgeFactor: undefined,