+++ src/common/utxobased/info/litecoin.ts
@@ -2,6 +2,7 @@ import { EdgeCurrencyInfo } from 'edge-core-js/types'
 
 import { IMAGE_SERVER_URL } from '../../constants'
 import { CoinInfo, EngineInfo, PluginInfo } from '../../plugin/types'
+import { maximumFeeRateCalculator } from '../../plugin/util/maximumFeeRateCalculator'
 import { memoInfo } from './commonInfo'
 
 export const currencyInfo: EdgeCurrencyInfo = {
@@ -55,7 +56,7 @@ export const engineInfo: EngineInfo = {
   gapLimit: 10,
   defaultFee: 50000,
   feeUpdateInterval: 60000,
-  simpleFeeSettings: {
+  defaultFeeInfo: {
     lowFeeFudgeFactor: undefined,
     standardFeeLowFudgeFactor: undefined,
     standardFeeHighFudgeFactor: undefined,
@@ -66,7 +67,8 @@ export const engineInfo: EngineInfo = {
     standardFeeLow: '150',
     standardFeeHigh: '200',
     standardFeeLowAmount: '20000000',
-    standardFeeHighAmount: '981000000'
+    standardFeeHighAmount: '981000000',
+    maximumFeeRate: maximumFeeRateCalculator(currencyInfo, 63.64)
   }
 }