+++ src/common/utxobased/info/zcoin.ts
@@ -2,6 +2,7 @@ import { EdgeCurrencyInfo } from 'edge-core-js/types'
 
 import { IMAGE_SERVER_URL } from '../../constants'
 import { CoinInfo, EngineInfo, PluginInfo } from '../../plugin/types'
+import { maximumFeeRateCalculator } from '../../plugin/util/maximumFeeRateCalculator'
 import { memoInfo } from './commonInfo'
 
 export const currencyInfo: EdgeCurrencyInfo = {
@@ -49,7 +50,7 @@ export const engineInfo: EngineInfo = {
   gapLimit: 10,
   defaultFee: 1000,
   feeUpdateInterval: 60000,
-  simpleFeeSettings: {
+  defaultFeeInfo: {
     lowFeeFudgeFactor: undefined,
     standardFeeLowFudgeFactor: undefined,
     standardFeeHighFudgeFactor: undefined,
@@ -60,7 +61,8 @@ export const engineInfo: EngineInfo = {
     standardFeeLow: '50',
     standardFeeHigh: '100',
     standardFeeLowAmount: '173200',
-    standardFeeHighAmount: '8670000'
+    standardFeeHighAmount: '8670000',
+    maximumFeeRate: maximumFeeRateCalculator(currencyInfo, 1.39)
   }
 }